reorganized into labelsync package
