+++ controller/nodelabel_controller.go
@@ -5,8 +5,6 @@ package controller
 
 import (
 	"context"
-	"errors"
-	"fmt"
 	"strings"
 	"sync"
 	"time"
@@ -22,11 +20,10 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
-	"github.com/Azure/go-autorest/autorest/to"
 	"github.com/Azure/node-label-operator/azure"
 	azrsrc "github.com/Azure/node-label-operator/azure/computeresource"
-	"github.com/Azure/node-label-operator/conversion"
-	"github.com/Azure/node-label-operator/conversion/options"
+	"github.com/Azure/node-label-operator/labelsync"
+	"github.com/Azure/node-label-operator/labelsync/options"
 )
 
 const (
@@ -130,6 +127,9 @@ func (r *ReconcileNodeLabel) Reconcile(req reconcile.Request) (reconcile.Result,
 // pass VMSS -> tags info and assign to nodes on VMs (unless node already has label)
 func (r *ReconcileNodeLabel) reconcileVMSS(namespacedName types.NamespacedName, provider *azure.Resource,
 	node *corev1.Node, configOptions *options.ConfigOptions) error {
+
+	log := r.Log.WithValues("node-label-operator", namespacedName)
+
 	vmss, err := azrsrc.NewVMSS(r.ctx, provider.SubscriptionID, provider.ResourceGroup, provider.ResourceName)
 	if err != nil {
 		return err
@@ -137,7 +137,7 @@ func (r *ReconcileNodeLabel) reconcileVMSS(namespacedName types.NamespacedName, 
 
 	if configOptions.SyncDirection == options.TwoWay || configOptions.SyncDirection == options.ARMToNode {
 		// only update if there are changes to labels
-		patch, err := r.applyTagsToNodes(namespacedName, *vmss, node, configOptions)
+		patch, err := labelsync.TagsToNodes(namespacedName, *vmss, node, configOptions, log, r.Recorder)
 		if err != nil {
 			return err
 		}
@@ -151,7 +151,7 @@ func (r *ReconcileNodeLabel) reconcileVMSS(namespacedName types.NamespacedName, 
 	// assign all labels on Node to VMSS, if not already there
 	if configOptions.SyncDirection == options.TwoWay || configOptions.SyncDirection == options.NodeToARM {
 		// only update if there are changes to labels
-		tags, err := r.applyLabelsToAzureResource(namespacedName, *vmss, node, configOptions)
+		tags, err := labelsync.LabelsToAzureResource(namespacedName, *vmss, node, configOptions, log, r.Recorder)
 		if err != nil {
 			return err
 		}
@@ -170,13 +170,16 @@ func (r *ReconcileNodeLabel) reconcileVMSS(namespacedName types.NamespacedName, 
 
 func (r *ReconcileNodeLabel) reconcileVMs(namespacedName types.NamespacedName, provider *azure.Resource,
 	node *corev1.Node, configOptions *options.ConfigOptions) error {
+
+	log := r.Log.WithValues("node-label-operator", namespacedName)
+
 	vm, err := azrsrc.NewVM(r.ctx, provider.SubscriptionID, provider.ResourceGroup, provider.ResourceName)
 	if err != nil {
 		return err
 	}
 
 	if configOptions.SyncDirection == options.TwoWay || configOptions.SyncDirection == options.ARMToNode {
-		patch, err := r.applyTagsToNodes(namespacedName, *vm, node, configOptions)
+		patch, err := labelsync.TagsToNodes(namespacedName, *vm, node, configOptions, log, r.Recorder)
 		if err != nil {
 			return err
 		}
@@ -188,7 +191,7 @@ func (r *ReconcileNodeLabel) reconcileVMs(namespacedName types.NamespacedName, p
 	}
 
 	if configOptions.SyncDirection == options.TwoWay || configOptions.SyncDirection == options.NodeToARM {
-		tags, err := r.applyLabelsToAzureResource(namespacedName, *vm, node, configOptions)
+		tags, err := labelsync.LabelsToAzureResource(namespacedName, *vm, node, configOptions, log, r.Recorder)
 		if err != nil {
 			return err
 		}
@@ -205,129 +208,9 @@ func (r *ReconcileNodeLabel) reconcileVMs(namespacedName types.NamespacedName, p
 	return nil
 }
 
-// return patch with new labels, if any, otherwise return nil for no new labels or an error
-func (r *ReconcileNodeLabel) applyTagsToNodes(namespacedName types.NamespacedName, computeResource azrsrc.ComputeResource, node *corev1.Node, configOptions *options.ConfigOptions) ([]byte, error) {
-	log := r.Log.WithValues("node-label-operator", namespacedName)
-
-	newLabels := map[string]*string{} // should allow for null JSON values
-	for tagName, tagVal := range computeResource.Tags() {
-		if !conversion.ValidLabelName(tagName) {
-			log.V(0).Info("invalid label name", "tag name", tagName)
-			continue
-		}
-		if !conversion.ValidLabelVal(*tagVal) {
-			log.V(0).Info("invalid label value", "tag value", *tagVal)
-			continue
-		}
-		validLabelName := conversion.ConvertTagNameToValidLabelName(tagName, configOptions.LabelPrefix)
-		labelVal, ok := node.Labels[validLabelName]
-		if !ok {
-			// add tag as label
-			log.V(1).Info("applying tags to nodes", "tag name", tagName, "tag value", *tagVal)
-			newLabels[validLabelName] = tagVal
-		} else if labelVal != *tagVal {
-			switch configOptions.ConflictPolicy {
-			case options.ARMPrecedence:
-				// set label anyway
-				log.V(1).Info("overriding existing node label with ARM tag", "tag name", tagName, "tag value", tagVal)
-				newLabels[validLabelName] = tagVal
-			case options.NodePrecedence:
-				// do nothing
-				log.V(0).Info("name->value conflict found", "node label value", labelVal, "ARM tag value", *tagVal)
-			case options.Ignore:
-				// raise k8s event
-				r.Recorder.Event(node, "Warning", "ConflictingTagLabelValues",
-					fmt.Sprintf("ARM tag was not applied to node because a different value for '%s' already exists (%s != %s).", tagName, *tagVal, labelVal))
-				log.V(0).Info("name->value conflict found, leaving unchanged", "label value", labelVal, "tag value", *tagVal)
-			default:
-				return nil, errors.New("unrecognized conflict policy")
-			}
-		}
-	}
-
-	// delete labels if tag has been deleted
-	// if conflict policy is node precedence (which it will most likely not be), then don't delete tags if they exist on node
-	if labelDeletionAllowed(configOptions) {
-		for labelFullName, labelVal := range node.Labels {
-			if conversion.HasLabelPrefix(labelFullName, configOptions.LabelPrefix) {
-				// check if exists on vm/vmss
-				labelName := conversion.LabelWithoutPrefix(labelFullName, configOptions.LabelPrefix)
-				_, ok := computeResource.Tags()[labelName]
-				if !ok { // if label doesn't exist on ARM resource, delete
-					log.V(1).Info("deleting label from node", "label name", labelFullName, "label value", labelVal)
-					delete(node.Labels, labelFullName) // for some reason this is needed
-					newLabels[labelFullName] = nil     // this should becomes 'null' in JSON, necessary for merge patch
-				}
-			}
-		}
-	}
-
-	if len(newLabels) == 0 { // to avoid unnecessary patching
-		return nil, nil
-	}
-
-	patch, err := LabelPatchWithDelete(newLabels)
-	if err != nil {
-		return nil, err
-	}
-
-	return patch, nil
-}
-
-func (r *ReconcileNodeLabel) applyLabelsToAzureResource(namespacedName types.NamespacedName, computeResource azrsrc.ComputeResource, node *corev1.Node, configOptions *options.ConfigOptions) (map[string]*string, error) {
-	log := r.Log.WithValues("node-label-operator", namespacedName)
-
-	if len(computeResource.Tags()) >= conversion.MaxNumTags {
-		log.V(0).Info("can't add any more tags", "number of tags", len(computeResource.Tags()))
-		return computeResource.Tags(), nil
-	}
-
-	newTags := map[string]*string{}
-	for labelName, labelVal := range node.Labels {
-		if !conversion.ValidTagName(labelName, configOptions.LabelPrefix) {
-			log.V(2).Info("invalid tag name", "label name", labelName)
-			continue
-		}
-		if !conversion.ValidTagVal(labelVal) {
-			log.V(2).Info("invalid tag name", "label name", labelName)
-			continue
-		}
-		validTagName := conversion.ConvertLabelNameToValidTagName(labelName, configOptions.LabelPrefix)
-		tagVal, ok := computeResource.Tags()[validTagName]
-		if !ok {
-			// add label as tag
-			log.V(1).Info("applying labels to Azure resource", "label name", labelName, "label value", labelVal)
-			newTags[validTagName] = to.StringPtr(labelVal)
-		} else if *tagVal != labelVal {
-			switch configOptions.ConflictPolicy {
-			case options.NodePrecedence:
-				// set tag anyway
-				log.V(1).Info("overriding existing ARM tag with node label", "label name", labelName, "label value", labelVal)
-				newTags[validTagName] = to.StringPtr(labelVal)
-			case options.ARMPrecedence:
-				// do nothing
-				log.V(0).Info("name->value conflict found", "node label value", labelVal, "ARM tag value", *tagVal)
-			case options.Ignore:
-				// raise k8s event
-				r.Recorder.Event(node, "Warning", "ConflictingTagLabelValues",
-					fmt.Sprintf("node label was not applied to Azure resource because a different value for '%s' already exists (%s != %s).", labelName, labelVal, *tagVal))
-				log.V(0).Info("name->value conflict found, leaving unchanged", "label value", labelVal, "tag value", *tagVal)
-			default:
-				return nil, errors.New("unrecognized conflict policy")
-			}
-		}
-	}
-
-	if len(newTags) == 0 { // if unchanged
-		return nil, nil
-	}
-
-	return newTags, nil
-}
-
 func (r *ReconcileNodeLabel) updateMinSyncPeriodLabels(node *corev1.Node) error {
 	r.lastUpdateLabel(node)
-	patch, err := LabelPatch(node.Labels)
+	patch, err := labelsync.LabelPatch(node.Labels)
 	if err != nil {
 		return err
 	}
