+++ labelsync/naming/naming.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-package conversion
+package naming
 
 import (
 	"fmt"
