+++ labelsync/naming/naming_test.go
@@ -1,4 +1,4 @@
-package conversion
+package naming
 
 import (
 	"fmt"
